Move sorter to utils and fetch data in hook
